+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/client_operation_documentation.rb
@@ -71,6 +71,7 @@ module AwsSdkCodeGenerator
     def option_tags(operation, api)
       if operation['input']
         shape = Api.shape(operation['input'], api)
+        return if shape['members'].nil?
         shape['members'].map do |member_name, member_ref|
           next if member_ref['documented'] === false
           docstring = Api.docstring(member_ref, api)
@@ -134,7 +135,7 @@ module AwsSdkCodeGenerator
           parts << "#\n"
           parts << "# @example Example: #{example['title']}\n#\n"
           if example['description'] && example['description'].length > 0
-            parts << "#{wrap_string(example['description'])}\n#\n"
+            parts << "#{Helper.wrap_string(example['description'], 120, "#   # ")}\n#\n"
           end
           parts += input.lines.map { |line| "#   " + line }
           if example['output']
@@ -190,13 +191,10 @@ module AwsSdkCodeGenerator
 
     def see_also_tag(operation, api)
       uid = api['metadata']['uid']
-      "# " + Crosslink.tag_string(uid, operation['name']) unless !Crosslink.taggable?(uid)
+      if api['metadata']['protocol'] != 'api-gateway' && Crosslink.taggable?(uid)
+        "# " + Crosslink.tag_string(uid, operation['name'])
+      end
     end
 
-    private
-
-    def wrap_string(content)
-      content.gsub(/(.{1,120})(\s+|\Z)/, "#   # \\1\n").chomp
-    end
   end
 end