+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/syntax_example_hash.rb
@@ -56,9 +56,11 @@ module AwsSdkCodeGenerator
 
     def struct(struct_shape, i, visited)
       lines = ['{']
-      struct_shape['members'].each_pair do |member_name, member_ref|
-        next if member_ref['documented'] === false
-        lines << struct_member(struct_shape, member_name, member_ref, i, visited)
+      unless struct_shape['members'].nil?
+        struct_shape['members'].each_pair do |member_name, member_ref|
+          next if member_ref['documented'] === false
+          lines << struct_member(struct_shape, member_name, member_ref, i, visited)
+        end
       end
       lines << "#{i}}"
       lines.join("\n")