+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/views/waiters_module.rb
@@ -5,6 +5,7 @@ module AwsSdkCodeGenerator
       def initialize(options)
         @module_name = options.fetch(:module_name)
         @waiters = Waiter.build_list(options.fetch(:waiters))
+        @custom = options.fetch(:custom)
       end
 
       # @return [String]
@@ -12,6 +13,12 @@ module AwsSdkCodeGenerator
 
       # @return [Array<Waiter>]
       attr_reader :waiters
+
+      # @return [String|nil]
+      def generated_src_warning
+        return if @custom
+        GENERATED_SRC_WARNING
+      end
 
     end
   end