+++ build_tools/aws-sdk-code-generator/spec/protocols_spec.rb
@@ -53,7 +53,7 @@ def client_for(suite, test_case, n)
   unless Aws.const_defined?(name)
     operation = test_case['given']
     operation['http'] ||= { 'method' => 'POST', 'requestUri' => '/' }
-    service = AwsSdkCodeGenerator::Service.new(
+    opts = {
       name: name,
       module_name: "Aws::#{name}",
       api: {
@@ -65,7 +65,11 @@ def client_for(suite, test_case, n)
       },
       gem_dependencies: { 'aws-sdk-core' => '3' },
       gem_version: '1.0.0',
-    )
+    }
+    if suite['metadata']['protocol'] == "api-gateway"
+      opts[:default_endpoint] = "https://foobar.us-west-2.amazonaws.com/test"
+    end
+    service = AwsSdkCodeGenerator::Service.new(opts)
     code = AwsSdkCodeGenerator::CodeBuilder.new(
       aws_sdk_core_lib_path: File.expand_path('../../../../gems/aws-sdk-core/lib/', __FILE__),
       service: service,
@@ -150,7 +154,12 @@ def match_req_body(group, suite, test_case, http_req, it)
     when 'rest-xml'
       body = normalize_xml(body)
       expected_body = normalize_xml(expected_body)
-    else raise "unsported protocol: `#{protocol}'"
+    when 'api-gateway'
+      if body[0] == '{'
+        body = Aws::Json.load(body)
+        expected_body = Aws::Json.load(expected_body)
+      end
+    else raise "unsupported protocol: `#{protocol}`"
     end
     it.expect(body).to(eq(expected_body))
   end